# [OCaml] Optimizations