Just couple of small optimizations:

- Initialize taggetPostCount array outside of a loop
- Calculate posts length just once outside of a loop