@jinyus 2023-11-25
2x faster

Main: `100.00 ms | 2.13 s | 22.17 s`

This PR:
```
Ocaml:

        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32768k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32640k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32640k
        Processing time (w/o IO): 77ms
        total: 0.12s memory: 32768k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32512k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32640k
        Processing time (w/o IO): 77ms
        total: 0.12s memory: 32768k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32640k
        Processing time (w/o IO): 78ms
        total: 0.13s memory: 32512k
        Processing time (w/o IO): 77ms
        total: 0.13s memory: 32640k

Ocaml:

        Processing time (w/o IO): 1263ms
        total: 1.52s memory: 103644k
        Processing time (w/o IO): 1216ms
        total: 1.47s memory: 103776k
        Processing time (w/o IO): 1283ms
        total: 1.54s memory: 103644k

Ocaml:

        Processing time (w/o IO): 10576ms
        total: 11.57s memory: 309896k
        Processing time (w/o IO): 10547ms
        total: 11.53s memory: 309892k
        Processing time (w/o IO): 10510ms
        total: 11.50s memory: 309764k

```